requirements.txt
flask
tensorflow
numpy
pillow
streamlit